+++ CHANGELOG.md
@@ -24,3 +24,5 @@ We are pleased to announce Helidon 4.0.0 a major release that includes significa
   - higher weight means a component is more important 
   - moved priority related types to MP config (as that is the lowest level MP module)
   - replaces all instances in SE that use priority with weight (no dependency on Jakarta, predictible and easy to understand behavior)
+- Introduction of `MediaType` as the abstraction of any media type, as used by Config, static content and HTTP in general. See `MediaType` and `MediaTypes`
+-